Add span parameter to footer sections
Allow having a footer section that spans multiple columns but does not
have multiple columns of list items. Useful if you have multiple rows of
sections and want to align a section with only a few items with a
section with many items (for instance the current GOV.UK footer layout).