wip! Use hack to use flexbox like grid
Based on [[1]]. We can't use grid autoplacement in any version of
Internet Explorer, so we have to go back to flexbox, but we use
border-box sizing and padding to make the gutter work the way we want.
(It turns out ironically that we do the same thing for our grids, but I
didn't realise this until later).

[1]: https://css-tricks.com/css-grid-in-ie-faking-an-auto-placement-grid-with-gaps/