Remove redundant 'navigation' from landmark name
In line with the section 'Composing Effective and User-friendly Accessible Names' in WAI-ARIA Authoring Practices 1.2 [1], remove the word navigation from the aria-label on the navigation.

'Do NOT include a WAI-ARIA role name in the accessible name. For example, do not include [..] the word "navigation" in the name of a navigation region. Doing so would create duplicate screen reader output since screen readers convey the role of an element in addition to its name.'

Without this change, assistive technologies report this as 'Navigation menu navigation'.

Also remove the word 'navigation' from the button that shows or hides the menu, as it also seems redundant – the word 'menu' should be descriptive enough for users to understand what it represents, especially when encountered inside a header landmark.

[1]: https://www.w3.org/TR/wai-aria-practices-1.2/#naming_effectively