Add helper function for generating module name
Add a new function to our helper-functions.js that
converts a component name to the Javascript module name.

This means we can use the function elsewhere, such as tests, without
duplicating the logic. It also makes sense to put together with the function
that generates the macro names, as the two are linked