Remove default font styles from cookie banner Sass
Removes the default font styles (setting the font and font-size) from
the cookie banner Sass.

This work was prompted by https://github.com/alphagov/govuk-design-system/issues/1886

With the default font styles present, it's very difficult to spot if you forget
to add the relevant `govuk-body` styles to any `html` you pass into the component.
We had missed this ourselves when we added the cookie banner to the design system.

This change should only affect users who are passing `html` to the cookie banner.